% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iris_tools.R
\name{print.iris_summary}
\alias{print.iris_summary}
\title{Print Method for iris_summary}
\usage{
\method{print}{iris_summary}(x, ...)
}
\arguments{
\item{x}{An object of class iris_summary.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print Method for iris_summary
}
